create rule for updating and deleting user in controller

+++ src/main/java/streetwalker/userservice/controllers/UserController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import streetwalker.userservice.dto.*;
 import streetwalker.userservice.models.User;
 import streetwalker.userservice.services.UserService;
+import streetwalker.userservice.services.security.SecurityUtils;
 
 import java.util.UUID;
 
@@ -19,14 +20,16 @@ import java.util.UUID;
 @RequestMapping("/users")
 public class UserController {
 
-    final UserService userService;
-
-    public UserController(UserService userService) {
+    private final UserService userService;
+    private final SecurityUtils securityUtils;
+    public UserController(UserService userService, SecurityUtils securityUtils) {
         this.userService = userService;
+        this.securityUtils = securityUtils;
     }
-
+    //переделать в возвращение dto
     @GetMapping
     public ResponseEntity<Page<User>> getAllUsers(Pageable pageable) {
+
         if(pageable != null) {
             return new ResponseEntity<>(userService.findAll(pageable), HttpStatus.OK);
         }
@@ -66,6 +69,11 @@ public class UserController {
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<UserDTO> update(@RequestBody @Validated UserUpdateDTO userData, @PathVariable Long id) {
+
+        if (!securityUtils.isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(null);
+        }
         UserDTO user;
         try {
             user = userService.update(userData, id);
@@ -97,6 +105,10 @@ public class UserController {
 
     @PostMapping("/{userId}/friends/{friendId}")
     public ResponseEntity<?> addFriend(@PathVariable Long userId, @PathVariable Long friendId) {
+        if (!securityUtils.isCurrentUser(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(null);
+        }
         try {
             userService.addFriend(userId, friendId);
             return new ResponseEntity<>("Friend added successfully", HttpStatus.OK);
@@ -107,6 +119,10 @@ public class UserController {
 
     @DeleteMapping("/{userId}/friends/{friendId}")
     public ResponseEntity<?> removeFriend(@PathVariable Long userId, @PathVariable Long friendId) {
+        if (!securityUtils.isCurrentUser(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(null);
+        }
         try {
             userService.removeFriend(userId, friendId);
             return new ResponseEntity<>("Friend removed successfully", HttpStatus.OK);
@@ -215,6 +231,10 @@ public class UserController {
 
     @PostMapping("/{id}/password/reset")
     public ResponseEntity<?> createPasswordResetLink(@PathVariable Long id) {
+        if (!securityUtils.isCurrentUser(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(null);
+        }
         try {
             userService.createNewPassLinkAndSendToMail(id);
             return new ResponseEntity<>("Password reset link sent to email", HttpStatus.OK);
